Also remove median/mean duplication, opportunistically moved also variance and standard_deviation to stats.hpp file
