+++ src/vcfinfosummarize.cpp
@@ -8,12 +8,13 @@
 */
 
 #include "Variant.h"
-#include "split.h"
+#include "stats.hpp"
 #include "Fasta.h"
 #include "gpatInfo.hpp"
 #include <getopt.h>
 #include <algorithm>
 #include <numeric>
+
 
 using namespace std;
 using namespace vcflib;
@@ -38,18 +39,6 @@ void printSummary(char** argv) {
     exit(0);
 }
 
-double median(vector<double> &v)
-{
-    size_t n = v.size() / 2;
-    nth_element(v.begin(), v.begin()+n, v.end());
-    return v[n];
-}
-
-double mean(vector<double> &v)
-{
-    double sum = accumulate(v.begin(), v.end(), 0.0);
-    return sum / v.size();
-}
 
 enum StatType { MEAN, MEDIAN, MIN, MAX };
